Move compute parts tests into its own module
